Move Owners2AzureSearch logic to Auxiliary2AzureSearch (#748)
Progress on https://github.com/NuGet/Engineering/issues/2978
