+++ NuGet.Services.Metadata.sln
@@ -46,8 +46,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGet.Services.SearchServic
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGet.Services.SearchService.Tests", "tests\NuGet.Services.SearchService.Tests\NuGet.Services.SearchService.Tests.csproj", "{F009209D-A663-45E1-87E8-158569A0F097}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGet.Jobs.Owners2AzureSearch", "src\NuGet.Jobs.Owners2AzureSearch\NuGet.Jobs.Owners2AzureSearch.csproj", "{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGet.Jobs.Auxiliary2AzureSearch", "src\NuGet.Jobs.Auxiliary2AzureSearch\NuGet.Jobs.Auxiliary2AzureSearch.csproj", "{7E6903A4-DBE1-444E-A8E3-C1DBB58243E0}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B2B65510-309B-45F1-AA18-8E4066BFEB92}"
@@ -257,18 +255,6 @@ Global
 		{F009209D-A663-45E1-87E8-158569A0F097}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{F009209D-A663-45E1-87E8-158569A0F097}.Release|x64.ActiveCfg = Release|Any CPU
 		{F009209D-A663-45E1-87E8-158569A0F097}.Release|x64.Build.0 = Release|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Debug|x64.Build.0 = Debug|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Release|x64.ActiveCfg = Release|Any CPU
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB}.Release|x64.Build.0 = Release|Any CPU
 		{7E6903A4-DBE1-444E-A8E3-C1DBB58243E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7E6903A4-DBE1-444E-A8E3-C1DBB58243E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7E6903A4-DBE1-444E-A8E3-C1DBB58243E0}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -373,7 +359,6 @@ Global
 		{1F3BC053-796C-4A35-88F4-955A0F142197} = {F1C83FD9-A498-483E-ADFA-B55D82A14965}
 		{DD089AB9-6AB3-4ACA-8D63-C95A7935B2A7} = {97E23323-BA7A-48F0-A578-858B82B6D8FB}
 		{F009209D-A663-45E1-87E8-158569A0F097} = {F1C83FD9-A498-483E-ADFA-B55D82A14965}
-		{F81A7CA6-97D4-4958-A9C1-FF94FCA283CB} = {C86C6DEE-84E1-4E4E-8868-6755D7A8E0E4}
 		{7E6903A4-DBE1-444E-A8E3-C1DBB58243E0} = {C86C6DEE-84E1-4E4E-8868-6755D7A8E0E4}
 		{5ABE8807-2209-4948-9FC5-1980A507C47A} = {C86C6DEE-84E1-4E4E-8868-6755D7A8E0E4}
 		{C3F9A738-9759-4B2B-A50D-6507B28A659B} = {5DE01C58-D5F7-482F-8256-A8333064384C}
