+++ build.ps1
@@ -83,7 +83,6 @@ Invoke-BuildStep 'Set version metadata in AssemblyInfo.cs' {
             "src\NuGet.Services.AzureSearch\Properties\AssemblyInfo.g.cs", `
             "src\NuGet.Jobs.Db2AzureSearch\Properties\AssemblyInfo.g.cs", `
             "src\NuGet.Jobs.Catalog2AzureSearch\Properties\AssemblyInfo.g.cs", `
-            "src\NuGet.Jobs.Owners2AzureSearch\Properties\AssemblyInfo.g.cs", `
             "src\NuGet.Services.SearchService\Properties\AssemblyInfo.g.cs", `
             "src\NuGet.Jobs.Auxiliary2AzureSearch\Properties\AssemblyInfo.g.cs", `
             "src\NuGet.Jobs.Catalog2Registration\Properties\AssemblyInfo.g.cs", `
@@ -136,7 +135,6 @@ Invoke-BuildStep 'Creating artifacts' {
             "src\Ng\Ng.Operations.nuspec", `
             "src\NuGet.Jobs.Db2AzureSearch\NuGet.Jobs.Db2AzureSearch.nuspec", `
             "src\NuGet.Jobs.Catalog2AzureSearch\NuGet.Jobs.Catalog2AzureSearch.nuspec", `
-            "src\NuGet.Jobs.Owners2AzureSearch\NuGet.Jobs.Owners2AzureSearch.nuspec", `
             "src\NuGet.Jobs.Auxiliary2AzureSearch\NuGet.Jobs.Auxiliary2AzureSearch.nuspec", `
             "src\NuGet.Jobs.Catalog2Registration\NuGet.Jobs.Catalog2Registration.nuspec", `
             "src\NuGet.Jobs.RegistrationComparer\NuGet.Jobs.RegistrationComparer.nuspec"
