--- src/NuGet.Jobs.Owners2AzureSearch/NuGet.Jobs.Owners2AzureSearch.nuspec
@@ -1,16 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <metadata>
-    <id>Owners2AzureSearch</id>
-    <version>$version$</version>
-    <authors>.NET Foundation</authors>
-    <owners>.NET Foundation</owners>
-    <description>Owners2AzureSearch</description>
-    <copyright>Copyright .NET Foundation</copyright>
-  </metadata>
-  <files>
-    <file src="bin\$configuration$\*.*" target="bin" />
-    <file src="Scripts\*.ps1" />
-    <file src="Scripts\nssm.exe" />
-  </files>
-</package>
