--- src/NuGet.Jobs.Owners2AzureSearch/Properties/AssemblyInfo.cs
@@ -1,9 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("NuGet.Jobs.Owners2AzureSearch")]
-[assembly: ComVisible(false)]
-[assembly: Guid("f81a7ca6-97d4-4958-a9c1-ff94fca283cb")]
