--- src/NuGet.Jobs.Owners2AzureSearch/Scripts/PreDeploy.ps1
@@ -1,11 +1,0 @@
-. .\Functions.ps1
-
-$jobsToInstall = $OctopusParameters["Jobs.ServiceNames"].Split("{,}")
-
-Write-Host Removing services...
-
-$jobsToInstall.Split("{;}") | %{
-	Uninstall-NuGetService -ServiceName $_
-}
-
-Write-Host Removed services.
