+++ src/NuGet.Services.AzureSearch/Auxiliary2AzureSearch/UpdateOwnersCommand.cs
@@ -11,9 +11,9 @@ using Microsoft.Extensions.Options;
 using NuGet.Services.AzureSearch.AuxiliaryFiles;
 using NuGet.Services.Metadata.Catalog.Helpers;
 
-namespace NuGet.Services.AzureSearch.Owners2AzureSearch
+namespace NuGet.Services.AzureSearch.Auxiliary2AzureSearch
 {
-    public class Owners2AzureSearchCommand : IAzureSearchCommand
+    public class UpdateOwnersCommand : IAzureSearchCommand
     {
         private readonly IDatabaseAuxiliaryDataFetcher _databaseFetcher;
         private readonly IOwnerDataClient _ownerDataClient;
@@ -23,9 +23,9 @@ namespace NuGet.Services.AzureSearch.Owners2AzureSearch
         private readonly Func<IBatchPusher> _batchPusherFactory;
         private readonly IOptionsSnapshot<AzureSearchJobConfiguration> _options;
         private readonly IAzureSearchTelemetryService _telemetryService;
-        private readonly ILogger<Owners2AzureSearchCommand> _logger;
+        private readonly ILogger<UpdateOwnersCommand> _logger;
 
-        public Owners2AzureSearchCommand(
+        public UpdateOwnersCommand(
             IDatabaseAuxiliaryDataFetcher databaseFetcher,
             IOwnerDataClient ownerDataClient,
             IOwnerSetComparer ownerSetComparer,
@@ -34,7 +34,7 @@ namespace NuGet.Services.AzureSearch.Owners2AzureSearch
             Func<IBatchPusher> batchPusherFactory,
             IOptionsSnapshot<AzureSearchJobConfiguration> options,
             IAzureSearchTelemetryService telemetryService,
-            ILogger<Owners2AzureSearchCommand> logger)
+            ILogger<UpdateOwnersCommand> logger)
         {
             _databaseFetcher = databaseFetcher ?? throw new ArgumentNullException(nameof(databaseFetcher));
             _ownerDataClient = ownerDataClient ?? throw new ArgumentNullException(nameof(ownerDataClient));
@@ -95,7 +95,7 @@ namespace NuGet.Services.AzureSearch.Owners2AzureSearch
             finally
             {
                 stopwatch.Stop();
-                _telemetryService.TrackOwners2AzureSearchCompleted(outcome, stopwatch.Elapsed);
+                _telemetryService.TrackUpdateOwnersCompleted(outcome, stopwatch.Elapsed);
             }
         }
