+++ tests/NuGet.Services.AzureSearch.Tests/Auxiliary2AzureSearch/OwnerSetComparerFacts.cs
@@ -7,7 +7,7 @@ using Moq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NuGet.Services.AzureSearch.Owners2AzureSearch
+namespace NuGet.Services.AzureSearch.Auxiliary2AzureSearch
 {
     public class OwnerSetComparerFacts
     {
